Popups refactor (#449)
* Popups refactor
* Add title to USB modes popup
* cleanup

+++ radio/src/targets/flysky/buzzer_driver.cpp
@@ -331,7 +331,7 @@ void buzzerHeartbeat()
 
       if (buzzerState.tone.freqIncr) {
         uint32_t freqChange = BUZZER_BUFFER_DURATION * buzzerState.tone.freqIncr;
-        buzzerState.freq += limit<uint16_t>(BEEP_MIN_FREQ, freqChange, BEEP_MAX_FREQ);
+        buzzerState.freq += limit<uint32_t>(BEEP_MIN_FREQ, freqChange, BEEP_MAX_FREQ);
 
         buzzerOn(buzzerState.freq, g_eeGeneral.beepVolume);
       }
